refactor(conversation): split turn checkpoint analytics family
